requirements.txt
discord.py
ics
requests
arrow